[REFACTOR][mumi]: MailService 인터페이스/Impl 병합
