+++ src/main/java/kkukmoa/kkukmoa/user/service/MailService.java
@@ -1,5 +1,28 @@
 package kkukmoa.kkukmoa.user.service;
 
-public interface MailService {
-    void sendOtp(String to, String code);
+import lombok.RequiredArgsConstructor;
+import org.springframework.mail.SimpleMailMessage;
+import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class MailService {
+
+    private final JavaMailSender mailSender;
+
+    public void sendOtp(String to, String code) {
+        SimpleMailMessage msg = new SimpleMailMessage();
+        msg.setTo(to);
+        msg.setSubject("회원가입 인증번호 안내");
+        msg.setText("""
+                안녕하세요. 꾹모아입니다.
+                아래 인증번호를 입력해 주세요.
+
+                인증번호: %s
+
+                인증번호는 5분 후 만료됩니다.
+                """.formatted(code));
+        mailSender.send(msg);
+    }
 }
